Reverse idea of abstracting the text measurement.
The DOM stream assumes HTML throughout and although it does work in practice building SVGs the types are no longer strictly compatible, so I've decided not to use DOM Stream for the SVGs. Maybe a fully compatible SVG DOM Stream implementation could be built but it will take some thought.